Fixed bad queries causing crashes or false connection drops

+++ src/MusicDownloader/utils/io/install.java
@@ -1,6 +1,7 @@
 package MusicDownloader.utils.io;
 
 import MusicDownloader.utils.app.debug;
+import MusicDownloader.utils.app.resources;
 import org.apache.commons.io.FileUtils;
 import org.apache.commons.io.FilenameUtils;
 import org.jsoup.Jsoup;
@@ -8,64 +9,67 @@ import org.jsoup.Jsoup;
 import java.io.File;
 import java.io.IOException;
 import java.net.URL;
-import java.nio.file.FileSystems;
-import java.nio.file.Files;
-import java.nio.file.Path;
-import java.nio.file.Paths;
+import java.nio.file.*;
 
 public class install {
 
-    public static final String YOUTUBE_DL_SOURCE = "https://youtube-dl.org/downloads/latest/youtube-dl.exe";
     public static final String FFMPEG_SOURCE = "https://ffmpeg.zeranoe.com/builds/win64/static/";
+
+    private static final String os = System.getProperty("os.name").toLowerCase();
 
     public static boolean getYoutubeDl() throws IOException {
 
-        // Windows only as of now
-        if (!System.getProperty("os.name").startsWith("Windows")) {
-            debug.warn("Youtube-DL installation only supported on windows currently.");
-            return false;
-        }
+        if (os.contains("win")) {
+            FileUtils.copyURLToFile(
+                    new URL("https://youtube-dl.org/downloads/latest/youtube-dl.exe"),
+                    new File(resources.getInstance().getYoutubeDlExecutable())
+            );
 
-        // Need to check OS to determine where to install it and to install
-        FileUtils.copyURLToFile(
-                new URL(YOUTUBE_DL_SOURCE),
-                new File(System.getenv("ProgramFiles(X86)") + "/youtube-dl/youtube-dl.exe")
-        );
+        } else if (os.contains("mac")) {
+            // Run console commands to install, doesn't need admin
 
-        // Should catch network exception
+        } else if (os.contains("nix") || os.contains("nux") || os.contains("aix")) {
+            debug.warn("Youtube-DL install not currently supported on Linux.");
 
-        // Testing executable to confirm installation
-        return verifyExecutable("youtube-dl");
+        } else debug.warn("Install is not supported on this operating system.");
 
+        return verifyExecutable(resources.getInstance().getYoutubeDlExecutable());
     }
 
     public static boolean getFFMPEG() throws IOException {
-        String packageVersion = Jsoup.connect(FFMPEG_SOURCE).get().select("a").get(1).attr("href");
 
-        // Downloading zip
-        FileUtils.copyURLToFile(
-                new URL(FFMPEG_SOURCE + packageVersion),
-                new File(System.getenv("ProgramFiles(X86)") + "/youtube-dl/ffmpeg.zip")
-        );
+        if (os.contains("win")) {
+            String packageVersion = Jsoup.connect(FFMPEG_SOURCE).get().select("a").get(1).attr("href");
 
-        // Unzipping and extract relevant files
-        Path zipFile = Paths.get(System.getenv("ProgramFiles(X86)") + "/youtube-dl/ffmpeg.zip");
+            FileUtils.copyURLToFile(
+                    new URL(FFMPEG_SOURCE + packageVersion),
+                    new File(System.getenv("ProgramFiles(X86)") + "/youtube-dl/ffmpeg.zip")
+            );
 
-        for (String file: new String[]{"ffmpeg.exe", "ffplay.exe", "ffprobe.exe"})
-            try (java.nio.file.FileSystem fileSystem = FileSystems.newFileSystem(zipFile, null)) {
-                Files.copy(
-                        fileSystem.getPath(FilenameUtils.removeExtension(packageVersion) + "/bin/" + file),
-                        Paths.get(System.getenv("ProgramFiles(X86)") + "/youtube-dl/" + file)
-                );
-            }
+            // Unzipping and extract relevant files
+            for (String file : new String[]{"ffmpeg.exe", "ffplay.exe", "ffprobe.exe"})
+                try (FileSystem fileSystem = FileSystems.newFileSystem(Paths.get(System.getenv("ProgramFiles(X86)") + "/youtube-dl/ffmpeg.zip"), null)) {
+                    Files.copy(
+                            fileSystem.getPath(FilenameUtils.removeExtension(packageVersion) + "/bin/" + file),
+                            Paths.get(System.getenv("ProgramFiles(X86)") + "/youtube-dl/" + file)
+                    );
+                }
 
-        // Delete irrelevant files
-        if (!new File(System.getenv("ProgramFiles(X86)") + "/youtube-dl/ffmpeg.zip").delete())
-            debug.warn("Failed to delete: " + System.getenv("ProgramFiles(X86)") + "/youtube-dl/ffmpeg.zip");
+            if (!new File(System.getenv("ProgramFiles(X86)") + "/youtube-dl/ffmpeg.zip").delete())
+                debug.warn("Failed to delete: " + System.getenv("ProgramFiles(X86)") + "/youtube-dl/ffmpeg.zip");
+
+        } else if (os.contains("mac")) {
+
+            // Install homebrew install ffmpeg
 
 
-        // Test
-        return verifyExecutable("ffmpeg");
+        } else if (os.contains("nix") || os.contains("nux") || os.contains("aix")) {
+
+            // Linux
+
+        } else debug.warn("Install is not supported on this operating system.");
+
+        return verifyExecutable(resources.getInstance().getFfmpegExecutable());
 
     }
